ViewError: mandatory body and Python 2.4 compat
Make the body of the ViewError exception mandatory, since many HTTP
error codes require an exception.

Also take the opportunity to add a bit of Python 2.4 compatibility:
exceptions are not new-style classes yet, so don't use super().

Finally, bump the minor version, since we changed the signature of
the ViewError initialization.